+++ src/main/java/techcourse/myblog/controller/CommentController.java
@@ -1,7 +1,5 @@
 package techcourse.myblog.controller;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -17,9 +15,6 @@ import techcourse.myblog.service.CommentService;
 @SessionAttributes("user")
 @Controller
 public class CommentController {
-
-    private static final Logger log = LoggerFactory.getLogger(CommentController.class);
-
     private final CommentService commentService;
     private final ArticleService articleService;
 
@@ -45,12 +40,10 @@ public class CommentController {
     }
 
     @DeleteMapping("/{commentId}")
-    private String deleteComment(@PathVariable Long commentId) {
-        log.error("asdfid, {}", commentId);
+    private String deleteComment(@PathVariable Long commentId, @ModelAttribute User user) {
+        checkOwner(commentId, user);
         Comment comment = commentService.findById(commentId);
-        log.error("asdfid1, {}", commentId);
         Long articleId = comment.getArticle().getId();
-        log.error("asdfid2, {}", commentId);
         commentService.delete(commentId);
 
         return "redirect:/articles/" + articleId;
@@ -68,5 +61,4 @@ public class CommentController {
             throw new MisMatchAuthorException("댓글을 작성한 유저만 수정할 수 있습니다.");
         }
     }
-
 }
